requirements.txt
altgraph==0.17.4
packaging==25.0
pefile==2023.2.7
pillow==10.4.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pywin32-ctypes==0.2.3
setuptools==80.8.0
ttkbootstrap==1.13.8
